# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quinn AI Personal Assistant API
  description: |
    Quinn is an AI-powered personal assistant that helps manage emails, banking, 
    food planning, life tasks, and more. All endpoints require authentication via 
    AWS Cognito JWT tokens.
  version: 1.0.0
  contact:
    name: Quinn Support
    email: support@quinn.ai
servers:
  - url: https://8aqsagpkp6.execute-api.eu-west-2.amazonaws.com/dev
    description: Development server
  - url: https://api.quinn.ai
    description: Production server (future)
security:
  - bearerAuth: []
...
paths:
...
  /calls:
    get:
      tags: [Calls]
      summary: Get call history
      responses:
        "200":
          description: Calls retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{callId}:
    get:
      tags: [Calls]
      summary: Get call by ID
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Unauthorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Resource not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Internal server error"
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
        code:
          type: string
...
    Call:
      type: object
      properties:
        callId:
          type: string
        userId:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [initiated, ringing, answered, completed, missed, failed]
        phoneNumber:
          type: string
        duration:
          type: integer
          description: Duration in seconds
        transcript:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time